requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
jinja2==3.1.4

# AI/ML
google-generativeai==0.8.3
chromadb==0.5.23
numpy<2.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Data Processing
pydantic==2.10.3
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1

